requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.29
aiohttp==3.9.5
flask==3.0.3
python-dotenv==1.0.1
